apps/python/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
asyncpg==0.29.0
pgvector==0.3.2
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2
pgvector==0.3.2

# Cache / Queue
redis==5.0.8

# AI APIs
anthropic==0.34.0
openai==1.45.0
google-generativeai==0.8.1

# Embeddings
sentence-transformers==3.1.1
numpy==1.26.4

# HTTP client
httpx==0.27.2

# Environment & config
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2

# Security / crypto
cryptography==43.0.1

# Utilities
python-multipart==0.0.9
